Death Explosion and added keybinds
Maybe some other changes I can't think of off the top of my head.
You can now press "z" for movement, "x" for attack, and spacebar to end
the turn.
There is a death explosion but it's giant and pink for some reason even
though it looked fine in the space shooter game.
Also i think if we just commit the files that start with "Assets" we can
avoid some problems with syncing.